feat: Add middleware to log request hostname prefix
This commit adds a new middleware to the Express application. The middleware logs the request's hostname prefix and allows it to proceed to the next handler. If the hostname does not contain a dot, it sends a 400 response with the message "Invalid hostname". The middleware is then used in the application to handle requests to the "/user" route.

The commit also includes the addition of the "express" dependency in the package.json file.

```